Rewrite DeltaExpr from Word to PVector
The new implementation is more generic, easier to maintain and
provides better type-safety. In my experiments it led to both
performance losses and gains, but in all cases the effect was small
(less than 10%). Performance loss is likely caused by the fact
that PVector occupies at least 24 bytes while Word is only 16 bytes.
If necessary, this can be addressed without the loss of generality
by replacing absl::InlinedVector with a custom implementation.